moving SpotLight to spot_light.rs
